1-FE: Move files into folders
