+++ Okane.Client/src/features/navigation/services/router.ts
@@ -6,10 +6,12 @@ import DashboardPage from '@/shared/pages/DashboardPage.vue'
 import LoginPage from '@/shared/pages/LoginPage.vue'
 import RegisterPage from '@/shared/pages/RegisterPage.vue'
 
+import { useAuthStore } from '@/features/auth/stores/useAuthStore'
+
 export enum RouteName {
-  DashboardPage =  'DashboardPage',
-  LoginPage = "LoginPage",
-  RegisterPage = "RegisterPage",
+  DashboardPage = 'DashboardPage',
+  LoginPage = 'LoginPage',
+  RegisterPage = 'RegisterPage',
 }
 
 export const ROUTES: Array<RouteRecordRaw> = [
@@ -22,17 +24,25 @@ export const ROUTES: Array<RouteRecordRaw> = [
     path: '/login',
     name: RouteName.LoginPage,
     component: LoginPage,
-    meta: { isPublic: true }
+    meta: { isPublic: true },
   },
   {
     path: '/register',
     name: RouteName.RegisterPage,
     component: RegisterPage,
-    meta: { isPublic: true }
-  }
+    meta: { isPublic: true },
+  },
 ] as const
 
 export const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: ROUTES,
-});
+})
+
+router.beforeEach((to) => {
+  const authStore = useAuthStore()
+
+  if (!to.meta.isPublic && !authStore.isLoggedIn) {
+    return { name: RouteName.LoginPage }
+  }
+})
