+++ Okane.Client/src/main.ts
@@ -6,12 +6,12 @@ import { createPinia } from 'pinia'
 // Internal
 import App from '@/App.vue'
 
-import { router } from '@/features/navigation/router';
+import { router } from '@/features/navigation/services/router'
 
 const app = createApp(App)
 
 app.use(createPinia())
-app.use(router);
+app.use(router)
 app.use(VueQueryPlugin)
 
 app.mount('#app')
